src/resources/initialize_lambda/requirements.txt
boto3==1.24.89 
psycopg2-binary==2.9.4
